mod: rearrange files + version and authors in __init__.py
